requirements.txt
# 기본 패키지
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
pandas==2.1.0
numpy==1.24.3

# 로깅 및 알림
colorlog==6.7.0
python-telegram-bot==13.15

# 데이터 분석 및 시각화
matplotlib==3.7.2
streamlit==1.27.0

# 데이터베이스
supabase==1.0.3

# 개발 도구
pytest==7.4.0
flake8==6.1.0
black==23.7.0
mypy==1.5.1 